// apps/web/src/generated/model/userResponseDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Core API
 * The core API for the platform
 * OpenAPI spec version: 1.0
 */

export interface UserResponseDto {
  /** Creation timestamp (ISO 8601) */
  createdAt: string;
  /** User email */
  email: string;
  /** Whether email is verified */
  emailVerified: boolean;
  /** User ID */
  id: string;
  /** Primary account ID */
  primaryAccountId?: string;
  /** User roles */
  roles: string[];
  /** Last update timestamp (ISO 8601) */
  updatedAt: string;
}